refactor: route cron subagent reads through registry seam
